+++ es-core/src/Settings.cpp
@@ -50,6 +50,7 @@ void Settings::setDefaults() {
     mBoolMap["DrawFramerate"] = false;
     mBoolMap["ShowExit"] = true;
     mBoolMap["Windowed"] = false;
+    mBoolMap["UseOSK"] = true;
 
 #ifdef _RPI_
 	// don't enable VSync by default on the Pi, since it already