+++ CHANGELOG.md
@@ -22,3 +22,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added retroachievements hardcore mode in game options
 - Added Creation of READY flag in /tmp directory
 - Added integer scale option in game options
+- Added a virtual keyboard